# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profiles": {
      "post": {
        "operationId": "ProfilesController_create",
        "summary": "Cria um novo profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "O profile foi criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "tags": ["Profile"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProfilesController_findAll",
        "summary": "Retorna todos os profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Retorna todos os dados do profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "tags": ["Profile"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/profiles/{id}": {
      "get": {
        "operationId": "ProfilesController_findOne",
        "summary": "Retorna o profile pelo ID informado",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do profile solicitado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "tags": ["Profile"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProfilesController_update",
        "summary": "Atualiza o profile pelo id informado",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dados do profile solicitado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileUpdate"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Acesso negado."
          }
        },
        "tags": ["Profile"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProfilesController_remove",
        "summary": "Remove o profile pelo ID informado",
        "parameters": [],
        "responses": {
          "200": {
            "description": "O profile, foi deletado com sucesso"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Acesso negado"
          },
          "404": {
            "description": "O profile não foi encontrado"
          }
        },
        "tags": ["Profile"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Pagamentos",
    "description": "API de pagamentos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token JWT"
      }
    },
    "schemas": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "José Silva"
          },
          "email": {
            "type": "string",
            "example": "jose.silva@email.com"
          },
          "document": {
            "type": "string",
            "example": "11122233344"
          },
          "document_type": {
            "type": "string",
            "enum": ["CPF", "CNPG"]
          },
          "user_type": {
            "type": "string",
            "enum": ["PF", "PJ"]
          },
          "phone": {
            "type": "string",
            "example": "31999999999"
          },
          "street": {
            "type": "string",
            "example": "Rua X"
          },
          "street_number": {
            "type": "string",
            "example": "10"
          },
          "complementary": {
            "type": "string",
            "example": "APTO 100"
          },
          "neighborhood": {
            "type": "string",
            "example": "Centro"
          },
          "city": {
            "type": "string",
            "example": "Belo Horizonte"
          },
          "state": {
            "type": "string",
            "example": "MG"
          },
          "zipcode": {
            "type": "string",
            "example": "30320320"
          },
          "country": {
            "type": "string",
            "example": "BR"
          }
        },
        "required": [
          "name",
          "email",
          "document",
          "document_type",
          "user_type",
          "phone",
          "street",
          "street_number",
          "complementary",
          "neighborhood",
          "city",
          "state",
          "zipcode"
        ]
      },
      "ProfileUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "José Silva"
          },
          "phone": {
            "type": "string",
            "example": "31999999999"
          },
          "street": {
            "type": "string",
            "example": "Rua X"
          },
          "street_number": {
            "type": "string",
            "example": "10"
          },
          "complementary": {
            "type": "string",
            "example": "APTO 100"
          },
          "neighborhood": {
            "type": "string",
            "example": "Centro"
          },
          "city": {
            "type": "string",
            "example": "Belo Horizonte"
          },
          "state": {
            "type": "string",
            "example": "MG"
          },
          "zipcode": {
            "type": "string",
            "example": "30320320"
          },
          "country": {
            "type": "string",
            "example": "BR"
          }
        },
        "required": [
          "name",
          "phone",
          "street",
          "street_number",
          "complementary",
          "neighborhood",
          "city",
          "state",
          "zipcode"
        ]
      },
...
    }
  }
}